Rename voteForm to stakeform
