+++ setup/jest.config.js
@@ -201,8 +201,8 @@ module.exports = {
     'src/modules/pos/validator/components/toggleIcon/index.js',
     'src/modules/common/components/amountField/index.js',
     'src/modules/pos/validator/hooks/useVoteAmountField.js',
-    'src/modules/pos/validator/components/VoteForm/VoteRow.js',
-    'src/modules/pos/validator/components/VoteForm/EmptyState.js',
+    'src/modules/pos/validator/components/StakeForm/VoteRow.js',
+    'src/modules/pos/validator/components/StakeForm/EmptyState.js',
     'src/modules/pos/validator/components/RegisterValidatorStatus/Status.js',
     'src/modules/pos/validator/components/UnlockBalanceForm/index.js',
     'src/modules/wallet/components/registerMultisigForm/memberField.js',
