+++ src/modules/pos/validator/components/StakeForm/stakeForm.css
